wip: Sentry setup
